# API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Goodemal Test API",
        "contact": "Alex Fredes",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Api rest para el test de goodmeal"
        }
    ],
    "paths": {
        "/api/productos": {
            "get": {
                "operationId": "1c686b20a4c57a8899e591139e274611",
                "responses": {
                    "200": {
                        "description": "listado de productos"
                    },
                    "500": {
                        "description": "error al listar las productos"
                    }
                }
            },
            "post": {
                "operationId": "585f59d443747aa6451e421f8703acac",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "producto_nombre": {
                                                "type": "string"
                                            },
                                            "producto_descripcion": {
                                                "type": "string"
                                            },
                                            "producto_precio": {
                                                "type": "string"
                                            },
                                            "producto_stock": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "producto_nombre": "Producto Nombre prueba",
                                    "producto_descripcion": "Descripcion Prueba",
                                    "producto_precio": 1000,
                                    "producto_stock": 10,
                                    "tienda_id": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "producto_nombre": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "producto_descripcion": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "producto_precio": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "producto_stock": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "tienda_id": {
                                            "type": "integer",
                                            "example": "content"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "operationId": "6607d74fb5390e185ce82d346e93bd6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "producto_nombre": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "producto_descripcion": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "producto_precio": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "producto_stock": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "tienda_id": {
                                            "type": "integer",
                                            "example": "content"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "70e51f43fa8a6479628f3d30d46bae42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "producto_nombre": {
                                                "type": "string"
                                            },
                                            "producto_descripcion": {
                                                "type": "string"
                                            },
                                            "producto_precio": {
                                                "type": "string"
                                            },
                                            "producto_stock": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "producto_nombre": "Producto Nombre prueba",
                                    "producto_descripcion": "Descripcion Prueba",
                                    "producto_precio": 1000,
                                    "producto_stock": 10,
                                    "tienda_id": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "producto_nombre": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "producto_descripcion": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "producto_precio": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "producto_stock": {
                                            "type": "string",
                                            "example": "content"
                                        },
                                        "tienda_id": {
                                            "type": "integer",
                                            "example": "content"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "7c670ef0d9ca1593d1c34e77d01b84a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Producto eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}